requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
everapi==0.1.1
freecurrencyapi==0.1.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pyaes==1.6.1
pyasn1==0.5.1
pyTelegramBotAPI==4.15.4
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
urllib3==2.2.0
yarl==1.9.4
